requirements.txt
# Web framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
alembic>=1.13.0
pgvector>=0.2.5

# Scheduling
apscheduler>=3.10.4

# HTTP & scraping
httpx>=0.27.0
beautifulsoup4>=4.12.0
feedparser>=6.0.11
praw>=7.7.1
pytrends>=4.9.2

# NLP & embeddings
sentence-transformers>=2.7.0
scikit-learn>=1.4.0
numpy>=1.26.0

# LLM
openai>=1.30.0
anthropicai>=0.26.0

# Utilities
python-dotenv>=1.0.1
pydantic>=2.7.0
rich>=13.7.0
